Move commit_md5sum_file and verify_checksum to checksum.rs
That is	a better logical place for those two functions and also
removes clutter from `fs.rs`.

Change-type: patch
